move cancel pop up to detail view, connect bindings
